requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
jinja2==3.1.6
xmlschema==3.4.2
lxml==5.3.0
pydantic==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.22
email-validator==2.3.0
pytest==9.0.2
mangum==0.21.0
boto3
pytest
sqlalchemy
passlib
python-jose
psycopg2-binary